Minor changes for better compilation.

+++ source/Tools/Explore/voronoi.hpp
@@ -878,7 +878,7 @@ namespace cinekine
     {
         printf("Arcs Remaining: %d\n", _arcCnt);
         printf("Circles Remaining: %d\n", _circleCnt);
-        printf("Edges alloced: %lu\n", _edges.size());
+        printf("Edges alloced: %d\n", (int) _edges.size());
     }
 
     float Fortune::leftBreakPoint(BeachArc* arc, float directrix)
